Cambio de tipo de variable Puntos Función de borrar relacion miembro-actividad

+++ src/main/java/com/albertsalud/members/model/entities/Activity.java
@@ -32,7 +32,7 @@ public class Activity {
 	private Date endDate;
 	
 	@Column(nullable = false)
-	private Long points;
+	private Float points;
 	
 	@Column(nullable = false)
 	private String code;
